Creating status helper
